Removed ViewController class
CL 249604, CL 249605
Removed the ViewController class. Moved the calls to MainPage into the
AppDelegate, and the Demo class header file imports. Changed the
storyboards to have MainPage as their custom class, and only referencing
outlet. Changed MainPage to be a UIViewController subclass.
Removed the setWebView function from MainPage.
Removed programmatic setting of the web view delegate and set it in the
storyboards instead.