CL 259609 Bug fixes
Updated to latest AndroidManifest.xml practice.
Fixed another bug where spaces in paths were not handled.
Added a check for FIXED gdEnableScript for iOS, which is no longer
needed.
Added a cordova platform update to fix a crash bug.
Added a refresh of the contributor plugin to fix a bug that the
Info.plist was overwritten.
Corrected a comment that was no conformant with official style.